moves datasource proxy tests into correct new pkg
